+++ package.json
@@ -19,6 +19,7 @@
     "react-native-elements": "^1.2.7",
     "react-native-gesture-handler": "^1.5.0",
     "react-native-image-picker": "^1.1.0",
+    "react-native-modal": "^11.5.3",
     "react-native-phone-input": "^0.2.4",
     "react-native-reanimated": "^1.4.0",
     "react-native-responsive-screen": "^1.3.0",
