+++ styles/comps/CameraRollStyles.js
@@ -1,0 +1,18 @@
+import {StyleSheet} from 'react-native';
+
+
+
+const CamRoll = StyleSheet.create({
+  Container: {
+    alignItems: 'center', justifyContent: 'center',
+    marginTop:20,
+  },
+
+  ProfPic: {
+    width: 120, height: 120, borderRadius:100,
+  }
+
+});
+
+
+export default CamRoll;
